Finished unit tests for new class and removed old classes.
